Moved tests from test_curve to elliptic_curve/element.rs
